refactor and improve timeline code
